86cyrhc1n: Move PropertyForm into PropertyFormAgent
