+++ app/client/ui-modules/property-form-agent/enum/PropertyFormMode.ts
